+++ src/app_core/mod.rs
@@ -2,6 +2,7 @@ use crate::{
     Library, key_handler::KeyBuffer, media_controls::MediaControlsHandle, player::PlayerHandle,
     ui_state::UiState,
 };
+use anyhow::Result;
 use crossbeam_channel::Receiver;
 use std::sync::Arc;
 
@@ -19,27 +20,8 @@ pub struct NoctaVox {
     pub(crate) ui: UiState,
     player: PlayerHandle,
     key_buffer: KeyBuffer,
-    library_refresh_rec: Option<Receiver<LibraryRefreshProgress>>,
+    library_refresh_rec: Option<Receiver<Result<Library>>>,
     media_controls: Option<MediaControlsHandle>,
     tick_sync: u32,
     restored_song_id: Option<u64>,
-}
-
-pub enum LibraryRefreshProgress {
-    Scanning {
-        progress: u8,
-    },
-    Processing {
-        progress: u8,
-        current: usize,
-        total: usize,
-    },
-    UpdatingDatabase {
-        progress: u8,
-    },
-    Rebuilding {
-        progress: u8,
-    },
-    Complete(crate::Library),
-    Error(String),
 }