+++ src/library/mod.rs
@@ -3,6 +3,7 @@ mod library;
 
 pub use domain::LEGAL_EXTENSION;
 pub use domain::{
-    Album, FileType, LongSong, Playlist, PlaylistSong, SimpleSong, SongDatabase, SongInfo,
+    Album, FileType, LongSong, Playlist, PlaylistSong, RefreshProgress, RefreshStage, SimpleSong,
+    SongDatabase, SongInfo,
 };
 pub use library::Library;