+++ src/ui_state/ui_state.rs
@@ -43,8 +43,7 @@ impl UiState {
             legal_songs: Vec::new(),
             playlists: Vec::new(),
 
-            library_refresh_progress: None,
-            library_refresh_detail: None,
+            library_refresh: None,
         }
     }
 }
@@ -127,28 +126,6 @@ impl UiState {
                 )
                 | (Pane::Popup, PopupType::Playlist(PlaylistAction::Rename))
         )
-    }
-}
-
-impl UiState {
-    pub fn set_library_refresh_progress(&mut self, progress: Option<u8>) {
-        self.library_refresh_progress = progress;
-    }
-
-    pub fn get_library_refresh_progress(&self) -> Option<u8> {
-        self.library_refresh_progress
-    }
-
-    pub fn set_library_refresh_detail(&mut self, detail: Option<String>) {
-        self.library_refresh_detail = detail;
-    }
-
-    pub fn get_library_refresh_detail(&self) -> Option<&str> {
-        self.library_refresh_detail.as_deref()
-    }
-
-    pub fn is_library_refreshing(&self) -> bool {
-        self.library_refresh_progress.is_some()
     }
 }